Experiment: Launch and configure Syncthing
This is an experiment where I launched and joined a synthing node to an
existing folder. The folder was paused and the idea was to use this
instance only as an API gateway. This did not work because the events
stopped as well... anyhow, it was a fun experiment that I check in if I
need this code for the future.

I the next commit I roll roll back to plan b with is:

* Use my external existing Syncthing instance to listen for events
* Mount the sync thing files read only inside this container to calc
  hashes